Доработал структуру проекта
