--- js/data/structure.js
@@ -1,166 +1,0 @@
-// Тип заданий
-export const TaskType = {
-  GUESS_TWO: `two-of-two`,
-  GUESS_ONE: `tinder-like`,
-  FIND: `one-of-three`
-};
-
-// Сопоставляем тип задания и заголовок названия
-const taskTypeToTaskTitle = {
-  [TaskType.GUESS_TWO]: `Угадайте для каждого изображения фото или рисунок?`,
-  [TaskType.GUESS_ONE]: `Угадайте для каждого изображения фото или рисунок?`,
-  [TaskType.FIND]: `Угадайте для каждого изображения фото или рисунок?`,
-};
-
-// Структура заданий
-export const TASKS = [
-  {
-    type: TaskType.GUESS_TWO,
-    title: taskTypeToTaskTitle[TaskType.GUESS_TWO],
-    questions: [
-      {
-        type: `photo`,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      },
-      {
-        type: `paint`,
-        img: `https://k42.kn3.net/CF42609C8.jpg`
-      }
-    ]
-  },
-  {
-    type: TaskType.GUESS_ONE,
-    title: taskTypeToTaskTitle[TaskType.GUESS_ONE],
-    questions: [
-      {
-        type: `photo`,
-        img: `https://i.imgur.com/DiHM5Zb.jpg`,
-      }
-    ]
-  },
-  {
-    type: TaskType.FIND,
-    title: taskTypeToTaskTitle[TaskType.FIND],
-    questions: [
-      {
-        type: `photo`,
-        isSelected: false,
-        img: `http://i.imgur.com/DKR1HtB.jpg`,
-      },
-      {
-        type: `paint`,
-        isSelected: true,
-        img: `https://k32.kn3.net/5C7060EC5.jpg`
-      },
-      {
-        type: `photo`,
-        isSelected: false,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      }
-    ]
-  },
-  {
-    type: TaskType.GUESS_TWO,
-    title: taskTypeToTaskTitle[TaskType.GUESS_TWO],
-    questions: [
-      {
-        type: `photo`,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      },
-      {
-        type: `paint`,
-        img: `https://k42.kn3.net/CF42609C8.jpg`
-      }
-    ]
-  },
-  {
-    type: TaskType.GUESS_ONE,
-    title: taskTypeToTaskTitle[TaskType.GUESS_ONE],
-    questions: [
-      {
-        type: `photo`,
-        img: `https://i.imgur.com/DiHM5Zb.jpg`,
-      }
-    ]
-  },
-  {
-    type: TaskType.FIND,
-    title: taskTypeToTaskTitle[TaskType.FIND],
-    questions: [
-      {
-        type: `photo`,
-        isSelected: false,
-        img: `http://i.imgur.com/DKR1HtB.jpg`,
-      },
-      {
-        type: `paint`,
-        isSelected: true,
-        img: `https://k32.kn3.net/5C7060EC5.jpg`
-      },
-      {
-        type: `photo`,
-        isSelected: false,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      }
-    ]
-  },
-  {
-    type: TaskType.GUESS_TWO,
-    title: taskTypeToTaskTitle[TaskType.GUESS_TWO],
-    questions: [
-      {
-        type: `photo`,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      },
-      {
-        type: `paint`,
-        img: `https://k42.kn3.net/CF42609C8.jpg`
-      }
-    ]
-  },
-  {
-    type: TaskType.GUESS_TWO,
-    title: taskTypeToTaskTitle[TaskType.GUESS_TWO],
-    questions: [
-      {
-        type: `photo`,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      },
-      {
-        type: `paint`,
-        img: `https://k42.kn3.net/CF42609C8.jpg`
-      }
-    ]
-  },
-  {
-    type: TaskType.GUESS_ONE,
-    title: taskTypeToTaskTitle[TaskType.GUESS_ONE],
-    questions: [
-      {
-        type: `photo`,
-        img: `https://i.imgur.com/DiHM5Zb.jpg`,
-      }
-    ]
-  },
-  {
-    type: TaskType.FIND,
-    title: taskTypeToTaskTitle[TaskType.FIND],
-    questions: [
-      {
-        type: `photo`,
-        isSelected: false,
-        img: `http://i.imgur.com/DKR1HtB.jpg`,
-      },
-      {
-        type: `paint`,
-        isSelected: true,
-        img: `https://k32.kn3.net/5C7060EC5.jpg`
-      },
-      {
-        type: `photo`,
-        isSelected: false,
-        img: `http://i.imgur.com/1KegWPz.jpg`,
-      }
-    ]
-  }
-];
