+++ js/modules/header/view.js
@@ -1,5 +1,5 @@
 import AbstractView from '../../abstract-view';
-import {Time, Life} from '../../data/game-params';
+import {Time, Life} from '../../settings';
 
 const drawHeart = (full) => {
   return `<img src='img/heart__${full ? `full` : `empty`}.svg' class='game__heart' alt='Life' width='32' height='32'>`;
