+++ js/modules/result/util.test.js
@@ -1,5 +1,5 @@
 import {assert} from 'chai';
-import {LIFE, answer, answers} from '../../data/mock';
+import {LIFE, answer, answers} from './mock';
 import {calculateAnswerScore, calculateTotalGameScore} from './util';
 
 describe(`функция calculateAnswerScore: проверка параметров`, () => {
