+++ js/timer/timer.test.js
@@ -1,5 +1,5 @@
 import {assert} from 'chai';
-import {createTimer} from './timer';
+import createTimer from './timer';
 
 describe(`Функция создания таймера`, () => {
